Split parts of queijo.cpp
