+++ cli/options.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Luau/Compiler.h"
+
+Luau::CompileOptions copts();
+
+bool getCodegenEnabled();
