+++ cli/require.h
@@ -1,0 +1,6 @@
+#pragma once
+
+struct lua_State;
+
+int lua_require(lua_State* L);
+int lua_requireFromSource(lua_State* L);
