+++ cli/spawn.h
@@ -1,0 +1,5 @@
+#pragma once
+
+struct lua_State;
+
+int lua_spawn(lua_State* L);
